Add database model constraints

+++ src/PurchaseOrderTracker.Domain/Models/SupplierAggregate/Product.cs
@@ -9,7 +9,7 @@ namespace PurchaseOrderTracker.Domain.Models.SupplierAggregate
         {
         }
 
-        public Product(string prodCode, string name, ProductCategory category, double price)
+        public Product(string prodCode, string name, ProductCategory category, decimal price)
         {
             ProdCode = prodCode ?? throw new ArgumentNullException(nameof(prodCode));
             Name = name ?? throw new ArgumentNullException(nameof(name));
@@ -21,7 +21,7 @@ namespace PurchaseOrderTracker.Domain.Models.SupplierAggregate
         public string ProdCode { get; set; }
         public string Name { get; set; }
         public ProductCategory Category { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public override string ToString()
         {
